server/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
breadability==0.1.20
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
dateparser==1.1.8
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
GoogleNews==1.6.8
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
Pillow==9.5.0
pycountry==22.3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
sumy==0.11.0
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
tzlocal==5.0.1
urllib3==2.0.2
vaderSentiment==3.3.2
Werkzeug==2.3.4
